requirements.txt
attrs==19.3.0
certifi==2019.11.28
dill==0.3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
typing==3.7.4.1
